backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
openai==1.10.0
python-dotenv==1.0.1
fpdf==1.7.2
pytest==8.0.0
httpx==0.26.0
google-generativeai
